Support invariant containers
Summary:
For issue https://github.com/facebook/pyrefly/issues/1799

Invariant containers have to be handled differently than the rest of the containers. There are two options I can think of

1- Do not expand the inner type of the container. This wil make lax mode more strict

2- Choose a covariant type as the output type to invariant container type. This option is more flexible but can accept containers that shouldn't be accepted

I went with option 2 for this solution because 1- It's more flexible and 2- The signature will look better on the IDE

If there are any other ideas people think I should consider, I'd be happy to look into those as well.

I tested the example list/set example on pydantic as well, and it works as expected. So Pydantic does provide this level of flexibility.

Reviewed By: yangdanny97

Differential Revision: D88768895

fbshipit-source-id: a2929229eee3c4206ff2617d6eaf20bdfa2c981d