Use iterable for most container types
Summary:
For some of the convariant containers, the resulting types involve invariant containers. For example, if we pass list[str] to a frozenset[str] field then list[str] won't assignable to list[LaxStr] due to invariance.

We can fix this by mapping those containers to `iterable` as well.

Again, this is only applicable to the signature of the `__init__` method. This does not affect how we read the field at all.

Reviewed By: stroxler

Differential Revision: D88773411

fbshipit-source-id: 333d315872909b5ec5b362640061ca0df8832364