error on break/continue outside of loop
Summary:
This diff sets up the proper plumbing for Ruff's SemanticSyntaxChecker, which was being used incorrectly before. This lets us emit errors for break/continue outside of loops.

Not all errors are emitted here because some are redundant w/ existing errors that Pyrefly emits, and to reduce the scope of changes in this diff. We can turn additional errors on in the future with appropriate testing.

closes https://github.com/facebook/pyrefly/issues/1796

Reviewed By: grievejia

Differential Revision: D88700199

fbshipit-source-id: ac7e2515af83f9fe3ae4dc260807a695712f989c