Include debug symbols in our release builds
Summary:
In issue 1790, we have a bug report of a stack overflow on a large, private
codebase. I believe that this is going to be a recurring problem in Pyrefly -
the codebases most likely to trigger crashes and infinite recursion are
large, complex codebases and are most likely found in enterprise adopters.

As a result, I think it is very important that we work on optimizing the flow
around reporting and debugging problems when we cannot necessarily pull the code;
the obvious way to get more information about a stack overflow (which in many
cases is guessable from just the stack) is to use a debugger like `lldb` to
find out what the repeating stack frames are.

But at the moment, we are shipping Pyrefly with no debug symbols, so unless
a user is ready to pull Pyrefly, set up a build environment, and build from
source they can't give us a good stack dump from their private codebase.

Debug symbols shouldn't cost perf, only binary size. From testing the builds,
it looks like the binary size goes from 19M to 20M with full debug info,
which I think is not much; this diff is even less substantial though, we use
lines only (`debug = 1` gives lines only, and at the moment I think the main
value is getting stack traces from end users)

Reviewed By: kinto0

Differential Revision: D88759681

fbshipit-source-id: 7617b7ba58d8d0519e10035a17faafb1d1a7c35d