Imlpement build system blocking functionality
Summary:
Adds a flag for `--build-system-blocking`, which causes any sourcedb rebuilds to block the main `LspQueue` thread. This is useful when doing benchmarking operations to see how long things take when the build system is being used, rather than possibly falling back to a filesystem fallback.

This diff just threads it through and calls the rebuild closure in `queue_source_db_rebuild_and_recheck` within the current thread (which should all be happening in the `LspQueue` after the previous diff, rather than sending it to the `sourcedb_queue`.

For the TSP server, the flag is not exposed, so we always pass `false`

Reviewed By: ndmitchell

Differential Revision: D88647606

fbshipit-source-id: 81963056b87dd9191384430d6e671c9341d9c399