Fix force requery when critical file changes
Summary:
A fix from earlier to cache sourcedb information across configs and config reloads caused non-Python file requeries to break. This diff fixes that by adding a `force` flag, which we pass to the sourcedb when handling `did_change_watched_files` and use to skip the short circuit if no new files are being queried.

If we switch to the better-cached build system query approach for Buck, we can get rid of this, so it's likely only a temporary fix.

Reviewed By: stroxler

Differential Revision: D88701833

fbshipit-source-id: 12641bf31903dedc8ca4ad56d1e5029b49c13060