Reload all configs with sourcedb
Summary:
Jia found that there can be a race condition if the following conditions happen:
1. we kick off the sourcedb query for files opened
2. a config change comes in, causing us to invalidate and produce new configs
3. sourcedb query returns, but only the old config is invalidated

We instead need to invalidate all configs associated with a given sourcedb instead of just the ones that were associated with it at the time that we initially kicked off the sourcedb query. This diff achieves that by keeping a mapping of sourcedbs to configs that hold that sourcedbs, and using that to get configs to invalidate find for.

Reviewed By: grievejia

Differential Revision: D88967779

fbshipit-source-id: 43266e884a06e87545f8c6e7b756380d2096ea42