Set debug level to `true` (full symbols)
Summary:
In D88759681, I originally did this and the binary size cost is about 5% - we have
a ~20MB binary and it costs about 1MB.

We downgraded to level 1 (which is supposed to be line-level only), which according
to docs ought to have worked for the intended use case of letting us do things
like connect a debugger to a stack overflow and get stack traces.

But this is not actually working - I confirmed that a `cargo build --release` on trunk
winds up giving only assembly level stack trace information. With `debug = true` (or
equivalently `debug = 2`, we get more than we really need (e.g. variable names) but
the most important thing is that we do indeed get a stack trace.

I think we should ship this now and possibly try to optimize later, 5% isn't a huge
cost given that we're a tool in beta and we're getting crash reports periodically.

Reviewed By: kinto0

Differential Revision: D89003462

fbshipit-source-id: 9788ab125078dc5ac361a293bc3c452652815c90