Don't strip debug symbols in maturin
Summary:
On top of the issue where lines didn't seem to be available from even a
raw `cargo build --release` (which has been addressed by setting the
`debug` setting in the release profile of the top level `Cargo.toml`,
I realized that maturin is separately stripping debug symbols.

Once again, the cost of debug symbols is only at about 5%, and our project
is in a state (new, and getting widespread use) where we are getting quite
a lot of crash reports.

So I think the value of being able to run in lldb or gdb and get a stack
trace is - at least for the moment - higher than that of minimizing the
binary.

We could look into shrinking the debug info and/or shipping a separate
debug table later if we want to optimize.

Reviewed By: ndmitchell, yangdanny97

Differential Revision: D89018460

fbshipit-source-id: 450ca861998226877fa5cda72bb035aab12aca89