Consider an uninitialised ClassVar on protocols as abstract (#1797)
Summary:
As per title, this diff addresses the last remaining issue in the
conformance test `protocols_explicit.py` by requiring declared classvar
members to be initialised for protocols.

The exception is that uninitialised class variables in interface files are not considered abstract. This exception is necessary to get around issues in typeshed.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1797

Test Plan: `python test.py`

Reviewed By: stroxler

Differential Revision: D88902254

Pulled By: yangdanny97

fbshipit-source-id: 049cc9ae9750107679d01e46fd8c00da7b44481b