Resolve django foreign key string and self references (#1822)
Summary:
Resolves ForeignKey models defined as `"self"` or a string forward ref.
This latter requires that the model is imported in the module (e.g. under `if TYPE_CHECKING`).

I'm unsure about how robust/clean this code is as my knowledge of this project (and Rust in general) is pretty limited at the moment.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1822

Test Plan: Removed expected failures for cases that work now and added a test with an imported model.

Reviewed By: stroxler

Differential Revision: D89154472

Pulled By: rchen152

fbshipit-source-id: 6f3533cc4bf84933a6510c0343aabd88f8fd684d