Fix panic caused by circular dep in TypeVar bound
Summary: Fixes https://github.com/facebook/pyrefly/issues/1659, a panic caused by a Var unexpectedly showing up where we expect a Quantified, caused by a cycle in class tparams. I couldn't find a good way to handle the cycle, so this diff instead avoids creating it in the first place by changing class binding creation to recognize that it does not need to create the binding that leads to the cycle. It's a pretty unsatisfactory solution, but it seems to work.

Reviewed By: yangdanny97

Differential Revision: D89236789

fbshipit-source-id: 37280df78835660945d121dd0f7c885b7dc32aed