Add recursion guard in `force_for_narrowing` (fix chrash in #1791)
Summary:
We have a reported crash in issue 1791 where a fuzzer-generated
snippet is able to generate a cycle in `Type::Var`s through
`Variable` pointers in the solver, which crashes `force_for_narrowing`.

The local fix is to introduce a recursion guard, and this is the fastest
way to address the problem. Note that since the behavior is unexpected,
I opt to raise a pyrefly internal error rather than silently have
poorly-defined behavior.

Patching the crash this way is straightforward, but I think it's actually not a
great solution - it's too easy to introduce this kind of bug if every
callsite nees a recurser guard. It would be much better if we could enforce
an invariant that these kinds of cycles never happen; I'm hoping to separately
see if we can instead detect this when the cycle is created

Fixes https://github.com/facebook/pyrefly/issues/1791

Facebook

NOTE: please don't land this for me. I'm still trying to get debug symbols flowing
all the way to pypi (had to first fix up our cargo build, then maturin) so that
we can debug some different crashes happening in the wild. It's really
helpful to have this small repro, and so far no users have run into it on real code,
only the fuzzer.

Reviewed By: yangdanny97

Differential Revision: D88691915

fbshipit-source-id: b107d05809ca2eb3f34a6a9e2fa2a61797f1b8e9