Add a `release-with-debug` profile
Summary:
This will allow users experiencing crashes to build Pyrefly from source and
debug with lldb to get usable stack traces. I've tried and failed to attach
debug symbols directly to the

We could experiment later with putting debug symbols off to the side in a
separate file, but that's significantly more involved since we'd have to work
out storing the additional artifacts, versioning schemes for how to correctly
fetch the file, and clear instructions on how to use it; I think for now a
release-with-debug profile means we're somewhat more prepared to engage with
bug reports like issue 1790.

Reviewed By: yangdanny97

Differential Revision: D89293690

fbshipit-source-id: 12eea2abb71722b1038a9078d6141fa90b896357