Add some more stress tests for issue 1706
Summary:
There was some doubt whether breaks and narrows are handled acceptably
in this implementation. These tests verify that we get pretty reasonable
behavior in all cases.

The only debatable behavior is that if a variable is assumed to exist in a
loop that definitely runs, we won't throw uninitialized local errors
downstream. But we always complain inside the loop, so the downstream errors
seem arguably duplicate / noisy anyway so I think this is acceptable.

Reviewed By: grievejia

Differential Revision: D89314182

fbshipit-source-id: ba448f3019d14acef8caa31b2b6e18c7534f4dc3