Simplify usage tracking
Summary:
We discovered recently that Pyrefly behaves badly, given our current
degree of cycle handling, if we try to be overly rigid about first-use
tracking in bindings; as a result we added the `narrowing_from` usage
kind so that narrows that happen inside a complex expression can inherit
the first-use-ness of the parent.

This works for the most part; it is probably exposed to nondeterminism,
but that exposure existed anyway and in practice we're not seeing problems.

This diff aims to simplify partial type inference a bit by leaning into
the existing approach more: let's not even make a new usage, instead we
can just treat the current idx as something that propagates down, so that
any time we begin an `ensure_expr` from a given idx we keep using that idx
even if work gets deferred into a separate binding.

Note: for the determinism issue, I'm hoping to work out a solver-time fix based
on storing temporary results in thread-local data structures that is both
simpler and has better guarantees of determinism than our current logic. This
fix is partially to make reasoning about the binding-time results a bit easier
in preparation for that change.

Reviewed By: rchen152

Differential Revision: D88413287

fbshipit-source-id: 5e879304ae153e5577fc269aef7076f0b3df2b53