Add some documentation around third party stub resolution.
Summary:
Since the logic is getting fairly complicated around third party stubs and how they are used, trying to add some documentation to the site which a decision tree showing how they are used.

I feel this tree might be confusing, so if anyone has any ideas for how to better represent this information happy to change this.

{F1984254606}

Reviewed By: yangdanny97

Differential Revision: D89395921

fbshipit-source-id: 6e951fe9234d8ae605f626cd281eb0f190e39e07