Resolve str() calls into __str__ or __repr__ method calls
Summary:
# Problem
During Pysa's analysis, we convert `str(x)` calls into `x.__str__()` or `x.__repr__()` calls depending on the type of `x`. This is done during a preprocessing step of the call graph building, which only relies on type information.
This does not currently work with pyrefly.

# Solution
Pyrefly will now add an extra call graph edge with origin `str-call-to-dunder-method`, which will be used to know how to preprocess the expression correctly in Pysa.

Reviewed By: tianhan0

Differential Revision: D89470110

fbshipit-source-id: 8d291122117b588a3bc9133d67c5e0879a409115