Add TelemetryServerState
Summary:
Adding new telemetry, I found it a bit confusing to query for sourcedb_available, because it wasn't clear why I was asking (except that I needed to pass it into new_dequeued).

This diff adds a wrapper type for this kind of "server state" data that we always log with telemetry.

Reviewed By: connernilsen

Differential Revision: D89595795

fbshipit-source-id: 1b17a775ef4fe384304ce6d2d8155f9b9625e7a9