Fix stack overflow in protocol -> forall -> protocol -> ...
Summary:
The previous diff added a test case tracking that Pyrefly infinitely recurses when
faced with a protocol that contains a Forall bound method that is compared to a
method of the protocol: inside a loop over the protocol, we'll repeatedly hit line
1556 and create fresh Vars unpacking the method (which is coming from an attribute
lookup, so it's the same method every time), and then proceed to do a comparison
that then (in the stack overflow case) itself tries to do the protocol comparison
again.

The problem here is that the protocol assumptions don't help, because the logic at
line 1556 (which is there to ensure that a forall method can match against anything,
either a forall or a concrete - so in itself the logic is not wrong) keeps making
fresh `Var` every time, which means we always wind up with a new protocol type
that never existed before and the full type equality check in protocol assumptions
never passes.

The fix is to treat a Forall almost as if it were itself a protocol, because this
is the point where we haven't yet inserted a fresh var: just add `(got, want)` to
our protocol assumptions list *before* we try to create fresh vars for `got` and
do the comparison. That way, if the comparison ever loops back to an identical
`(got, want)` pair recursively, we assume it's true.

I think they key thing here is that the logic in our recursive assumptions isn't actually
about protocols, it's about recursive types (this is very clear if you look at Pyre1's code,
which uses the exact same plumbing for both recursive aliases and protocols). The logic in
question is relevant to every potentially recursive type. And in the presence
of other recursive types, forall checks are potentially recursive. If we didn't create
fresh Vars this might not matter (we would catch the recursion elsewhere), but because
of the fresh Vars, the only way to handle it is to directly model recursion for
`forall` cases of `got`.

Fixes https://github.com/facebook/pyrefly/issues/1905

Reviewed By: yangdanny97

Differential Revision: D89604001

fbshipit-source-id: c735510f7da4cf15889793f30872cba5ec05b145