Forbid re-assignment of `Final` fields in constructor (#1920)
Summary:
Previously, the check for `Final` fields in classes did not check whether the final fields are provided with an initial value. This diff adds check for whether the final fields have been initialised in their declaration, and raises an error if they are re-assigned in constructors or other methods that are permitted to modify attributes.

Fixes one failing test in the conformance testsuite.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1920

Test Plan: `test.py`

Reviewed By: yangdanny97

Differential Revision: D89703283

Pulled By: rchen152

fbshipit-source-id: e0958ba44f87ffec7aaa07ad34790ecefe845b5e