Add BadDunderAll error kind for __all__ validation
Summary:
I promised a few improvements on top of D88489504. One of them is to use a separate error kind for the issue. Previously we were using `MissingModuleAttribute` for the purpose, but it doesn't feel right to me because `MissingModuleAttribute` is more about failures upon attribute lookup, and the issue we were detecting is more about missing `__all__` elements. I instead name it `BadDunderAll` to provide clearer semantics
and allowing users to configure this specific check independently.

Reviewed By: yangdanny97

Differential Revision: D89700988

fbshipit-source-id: 6462c53e7054cc951462f304e6e8c46e755a7459