Improve `__new__` special-casing for Self type
Summary: I was looking at D89479870 to try to understand why `object.__new__(...)` previously returned the wrong result, and I realized that the fix in that diff is incomplete: it emulates treating `Self` as a TypeVar for `object.__new__`, but we still do the wrong thing for custom `__new__` methods. This diff instead modifies `as_instance_attribute` and `as_class_attribute` to convert `Self` to a TypeVar when a `__new__` method is looked up. This is sort of ugly but gets us correct behavior (or at least behavior matching existing type checkers) in more cases. It also fixes an error in pytorch.

Reviewed By: yangdanny97

Differential Revision: D89743361

fbshipit-source-id: bcba305f12d1ad6459005efc99efe8148257cea6