use `for_display` for types used in error
Summary:
We forgot a spot (or it was newly added and slipped thru code review). I manually audited the rest of the usages of ErrorContext and TypeCheckKind, but other one-off error messages might still contain this.

This caused significant non-determinism in the mypy primer comments.

I think our existing nondet CI did not catch this because it only looks at error counts (tho I think due to flakiness it's frequently ignored ATM).

Ideally, there should be more safeguards than relying callers to process the type with the right method. Perhaps we should make types non-displayable - for example, make a wrapper `TypeForDisplay` struct that wraps a type, and make that the only way to pretty-print a type.

Reviewed By: rchen152

Differential Revision: D89854537

fbshipit-source-id: 2d67e3b9e455a7e03d1d9118247da3ebe5390eaa