Store the value of TYPE_CHECKING as a boolean in SysInfo
Summary:
Allows the value of `typing.TYPE_CHECKING` to be set to either true or false
when constructing a `SysInfo`. The main use case is to let other code analysis
projects use the `pyrefly_python` crate and prune `if TYPE_CHECKING` blocks.

Sets the default to `true` in `SysInfo::new`, so pyrefly code should not be
affected by this change.

Reviewed By: rchen152

Differential Revision: D89868143

fbshipit-source-id: e33182a71ea85e430deeeaaedf3dd7642d40c488