Pycharm: Support unopened file resolve for `provide_type` (#1918)
Summary:
Hello I am from Pycharm team and we want to support your perfect tool as first class citizen and interrate it to the type resolve pipeline as an option!

There was the issue for initial support for type resolve
https://github.com/facebook/pyrefly/issues/1181

It was perfectly but it works only for opened files, but we want to use this fucntion to provide types in libraries and python stubs.

I made I single suggestion how it can be supported.

Please check a PR and it will be great if it can be took to the main branch.
If it possible we want to add it ASAP to include the deep support of Pyrefly in Pycharm 2026.1 version!

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1918

Reviewed By: kinto0

Differential Revision: D89681826

Pulled By: yangdanny97

fbshipit-source-id: bfe71b323fe63a3b997d0db0283ced4f73f08812